% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mcad.R
\name{add_mcad}
\alias{add_mcad}
\title{Add MCAD variables for FY18}
\usage{
add_mcad(df)
}
\arguments{
\item{df}{data frame to add}
}
\description{
Add MCAD variables for FY18
}
\examples{
\dontrun{
df_mer <- add_mcad(df_mer) }
}
