% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pos.R
\name{add_pos}
\alias{add_pos}
\title{Add missing PMTCT_STAT_POS or TB_STAT_POS}
\usage{
add_pos(df, ind, pd = fy2019_targets)
}
\arguments{
\item{df}{dataframe with missing _POS}

\item{ind}{indicator, PMTCT_STAT or TB_STAT}

\item{pd}{period to add _POS to}
}
\description{
Add missing PMTCT_STAT_POS or TB_STAT_POS
}
