% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_netnew.R
\name{combine_netnew}
\alias{combine_netnew}
\title{Add net new to dataset}
\usage{
combine_netnew(df)
}
\arguments{
\item{df}{dataframe to create net new from and add it on}
}
\description{
Add net new to dataset
}
\examples{
\dontrun{
df_msd <- combine_netnew(df_msd)}


}
