% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative}
\alias{cumulative}
\title{Add cumulative value for current FY}
\usage{
cumulative(df, fy, qtr)
}
\arguments{
\item{df}{data frame to apply}

\item{fy}{current fiscual year}

\item{qtr}{current quarter}
}
\description{
Add cumulative value for current FY
}
\examples{
df_mer <- cumulative(cumulative, 2018, 1)
}
