% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export as base dataset(s) for PPR template}
\usage{
export(df, opunit, savepd)
}
\arguments{
\item{df}{dataframe to export}

\item{opunit}{which operatingunit should an output be created for?}

\item{savepd}{period to add to file name, eg FY2018Q1}
}
\description{
Export as base dataset(s) for PPR template
}
\examples{
\dontrun{
 export(df_mer, "Kenya", "FY2018Q1") }
 
}
