% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurefiller.R
\name{fill_future_pds}
\alias{fill_future_pds}
\title{Create/Fill future period columns with zero as placeholders}
\usage{
fill_future_pds(df, fy, qtr)
}
\arguments{
\item{df}{dataset to add future columns to}

\item{fy}{current fiscal year}

\item{qtr}{current quarter}
}
\description{
Create/Fill future period columns with zero as placeholders
}
\examples{
\dontrun{
df_mer <- fill_future_pds(df_mer, 2018, 1) }

}
