% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_keyinds.R
\name{filter_keyinds}
\alias{filter_keyinds}
\title{Subset dataframe to just relevant indicators}
\usage{
filter_keyinds(df, qtr)
}
\arguments{
\item{df}{dataframe to subset}

\item{qtr}{current quarter}
}
\description{
Subset dataframe to just relevant indicators
}
\examples{
df_ppr <- filter_keyinds(df_mer, 1)

}
