% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPPR.R
\name{genPPR}
\alias{genPPR}
\title{Generate PPR export data file}
\usage{
genPPR(filepath_msd, folderpath_output = "ExcelOutput",
  output_global = TRUE, output_ctry_all = TRUE,
  output_subset_type = NULL, ...)
}
\arguments{
\item{filepath_msd}{what is the file path to the MER Structured dataset? eg "~/ICPI/Data/MER_Structured_Dataset_OU_IM_FY17-19_20190517_v1_1.rds"}

\item{folderpath_output}{where do you want the output saved?, default = "ExcelOutput"}

\item{output_global}{export full dataset? logical, default = TRUE}

\item{output_ctry_all}{export each country? logicial, default = TRUE}

\item{output_subset_type}{select only subset, either "ou" or "mechid"}

\item{...}{add list of countries or mechanisms for `output_subset_type`, eg "18841", "14421"}
}
\description{
Purpose: generate output for Excel monitoring dashboard
NOTES
  - Data source: MER Structured Dataset PSNUxIM  [ICPI Data Store]
  - Report aggregates DSD and TA
  - Report looks at only Totals and coarse age (<15/15+) for testing and treatment
}
\examples{
\dontrun{
#export global file
  genPPR("~/ICPI/Data/MER_Structured_Dataset_OU_IM_FY17-19_20190517_v1_1.rds", output_ctry_all = FALSE)
#view global file
  df_ppr <- genPPR("~/ICPI/Data/MER_Structured_Dataset_OU_IM_FY17-19_20190517_v1_1.rds", output_global = FALSE, output_ctry_all = FALSE)
#export global and country specific files to populate PPR
  genPPR("~/ICPI/Data/MER_Structured_Dataset_OU_IM_FY17-19_20190517_v1_1.rds")
#export just Malawi and Kenya
  genPPR("~/ICPI/Data/MER_Structured_Dataset_OU_IM_FY17-19_20190517_v1_1.rds", output_global = FALSE, output_ctry_all = FALSE, folderpath_output = "ExcelOutput", output_subset_type = "ou", "Kenya", "Malawi")
#export two mechanims
  genPPR("~/ICPI/Data/MER_Structured_Dataset_OU_IM_FY17-19_20190517_v1_1.rds", output_global = FALSE, output_ctry_all = FALSE, folderpath_output = "ExcelOutput", output_subset_type = "mechid", "18234", "18544") }

}
