% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_oldtx.R
\name{import_oldtx}
\alias{import_oldtx}
\title{Import FY16Q4 data for TX_NET_NEW Calculation}
\usage{
import_oldtx(df, folderpath)
}
\arguments{
\item{df}{dataframe to add archived data onto}

\item{folderpath}{}
}
\description{
Import FY16Q4 data for TX_NET_NEW Calculation
}
