% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_nn_targets.R
\name{include_nn_targets}
\alias{include_nn_targets}
\title{Append NET_NEW target}
\usage{
include_nn_targets(df)
}
\arguments{
\item{df}{data frame to add NET_NEW Targets onto}
}
\description{
Append NET_NEW target
}
