% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_fldr.R
\name{initialize_fldr}
\alias{initialize_fldr}
\title{Initialize folder structure with sub folders}
\usage{
initialize_fldr(projectname, projectpath, ...)
}
\arguments{
\item{projectname}{name of the project}

\item{projectpath}{path to the folder where you want your project to sit}

\item{...}{subfolders to add in list, c("x", "y")}
}
\description{
Initialize folder structure with sub folders
}
\details{
Purpose: initialize packages, folder structure, and global file paths
Adapted from T. Essam, USAID [Stata]
}
\examples{
initialize_fldr("NewProject", "~/GitHub", "RawData", "Documents", "R", "Output")
}
