% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_keyinds.R
\name{key_ind}
\alias{key_ind}
\title{Identify indicators of interest}
\usage{
key_ind(qtr)
}
\arguments{
\item{qtr}{current quarter}
}
\description{
Identify indicators of interest
}
\examples{
\dontrun{
df_mer <- key_ind(1) }

}
