% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit.R
\name{limit}
\alias{limit}
\title{Limit}
\usage{
limit(df, fy)
}
\arguments{
\item{df}{data frame to apply}

\item{fy}{current fy, eg 2018}
}
\description{
Velcro identifies generic variable columns - last 6 pds, prior APR value, current cumulative 
  value, and current FY targets - and then limits output dataset to
}
\examples{
\dontrun{
df_output <- limit(df_mer, 2018)}
}
