% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggcheck.R
\name{pprcheck}
\alias{pprcheck}
\title{Check OU Aggregate MER Structured Dataset against PPR ouput from runPPR.R}
\usage{
pprcheck(filepath_fv, opunit)
}
\arguments{
\item{filepath_fv}{filepath to current PSNU_IM file}

\item{opunit}{operatingunit to check}
}
\description{
Check OU Aggregate MER Structured Dataset against PPR ouput from runPPR.R
}
\examples{
\dontrun{
check("ICPI/Data/ICPI_MER_Structured_Dataset_PSNU_IM_20180215_v1_3.Rds", "Malawi") }


}
