% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_impatt.R
\name{pull_impatt}
\alias{pull_impatt}
\title{Exract PLHIV and POP for PPR context}
\usage{
pull_impatt(folderpath)
}
\arguments{
\item{folderpath}{file path to the folder that contains the NAT_SUBNAT file}
}
\description{
Exract PLHIV and POP for PPR context
}
\examples{
\dontrun{
df_impatt <- pull_impatt("~/ICPI/Data") }

}
