% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprioritize.R
\name{reprioritize}
\alias{reprioritize}
\title{Change all prior prioritizations to what is set in current year}
\usage{
reprioritize(df)
}
\arguments{
\item{df}{dataframe to adjust prioritizations on}
}
\description{
Change all prior prioritizations to what is set in current year
}
