% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggcheck.R
\name{check}
\alias{check}
\title{Check OU Aggregate FV against PPR ouput from runPPR.R}
\usage{
check(filepath_fv, opunit)
}
\arguments{
\item{filepath_fv}{filepath to current PSNU_IM file}

\item{opunit}{operatingunit to check}
}
\description{
Check OU Aggregate FV against PPR ouput from runPPR.R
}
