% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netnew.R
\name{netnew}
\alias{netnew}
\title{Generate NET_NEW variable}
\usage{
netnew(df)
}
\arguments{
\item{df}{dataset to use/bind net new to}
}
\description{
Generate NET_NEW variable
}
\examples{
df_mer <- netnew(df_mer)

}
